Prototype Database and Draft Platform
This update includes a draft Event Management Platform which Includes a Backend connected to a prototype database enabling data to be passed from frontend to backend. Next step is to develop the database further and incorporate it into the website further while continuing to add features that take advantage of this connection.